Refactoring `UnitCellMatrices`
- moved into a separate file in framework (this is not LBS-specific)
- moved GeometryType into framework (not LBS-specific)
- removed duplicate comutations of unit cell matrices, all is done inside the UnitCellMatrices class
- moved SpatialWeightFunction into a separate file and removed its copies
