+++ modules/linear_boltzmann_solvers/lbs_problem/acceleration/diffusion_mip_solver.cc
@@ -7,6 +7,7 @@
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/math/spatial_discretization/finite_element/finite_element_data.h"
 #include "framework/math/spatial_discretization/spatial_discretization.h"
+#include "framework/math/unit_cell_matrices/unit_cell_matrices.h"
 #include "framework/math/functions/function.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
