+++ modules/linear_boltzmann_solvers/lbs_problem/acceleration/diffusion_pwlc_solver.cc
@@ -5,6 +5,7 @@
 #include "modules/linear_boltzmann_solvers/lbs_problem/acceleration/acceleration.h"
 #include "modules/linear_boltzmann_solvers/lbs_problem/lbs_structs.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
+#include "framework/math/unit_cell_matrices/unit_cell_matrices.h"
 #include "framework/math/spatial_discretization/spatial_discretization.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
