+++ modules/linear_boltzmann_solvers/lbs_problem/lbs_problem.cc
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "modules/linear_boltzmann_solvers/lbs_problem/lbs_problem.h"
+#include "framework/math/unit_cell_matrices/unit_cell_matrices.h"
 #include "modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/boundary/reflecting_boundary.h"
 #include "modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/boundary/vacuum_boundary.h"
 #include "modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/boundary/isotropic_boundary.h"
@@ -1019,124 +1020,17 @@ LBSProblem::ComputeUnitIntegrals()
   log.Log() << "Computing unit integrals.\n";
   const auto& sdm = *discretization_;
 
-  // Define spatial weighting functions
-  struct SpatialWeightFunction // SWF
-  {
-    virtual double operator()(const Vector3& pt) const { return 1.0; }
-    virtual ~SpatialWeightFunction() = default;
-  };
-
-  struct SphericalSWF : public SpatialWeightFunction
-  {
-    double operator()(const Vector3& pt) const override { return pt[2] * pt[2]; }
-  };
-
-  struct CylindricalSWF : public SpatialWeightFunction
-  {
-    double operator()(const Vector3& pt) const override { return pt[0]; }
-  };
-
-  auto swf_ptr = std::make_shared<SpatialWeightFunction>();
-  if (options_.geometry_type == GeometryType::ONED_SPHERICAL)
-    swf_ptr = std::make_shared<SphericalSWF>();
-  if (options_.geometry_type == GeometryType::TWOD_CYLINDRICAL)
-    swf_ptr = std::make_shared<CylindricalSWF>();
-
-  auto ComputeCellUnitIntegrals = [&sdm](const Cell& cell, const SpatialWeightFunction& swf)
-  {
-    const auto& cell_mapping = sdm.GetCellMapping(cell);
-    const size_t cell_num_faces = cell.faces.size();
-    const size_t cell_num_nodes = cell_mapping.GetNumNodes();
-    const auto fe_vol_data = cell_mapping.MakeVolumetricFiniteElementData();
-
-    DenseMatrix<double> IntV_gradshapeI_gradshapeJ(cell_num_nodes, cell_num_nodes, 0.0);
-    DenseMatrix<Vector3> IntV_shapeI_gradshapeJ(cell_num_nodes, cell_num_nodes);
-    DenseMatrix<double> IntV_shapeI_shapeJ(cell_num_nodes, cell_num_nodes, 0.0);
-    Vector<double> IntV_shapeI(cell_num_nodes, 0.);
-    std::vector<DenseMatrix<double>> IntS_shapeI_shapeJ(cell_num_faces);
-    std::vector<DenseMatrix<Vector3>> IntS_shapeI_gradshapeJ(cell_num_faces);
-    std::vector<Vector<double>> IntS_shapeI(cell_num_faces);
-
-    // Volume integrals
-    for (unsigned int i = 0; i < cell_num_nodes; ++i)
-    {
-      for (unsigned int j = 0; j < cell_num_nodes; ++j)
-      {
-        for (const auto& qp : fe_vol_data.GetQuadraturePointIndices())
-        {
-          IntV_gradshapeI_gradshapeJ(i, j) +=
-            swf(fe_vol_data.QPointXYZ(qp)) *
-            fe_vol_data.ShapeGrad(i, qp).Dot(fe_vol_data.ShapeGrad(j, qp)) *
-            fe_vol_data.JxW(qp); // K-matrix
-
-          IntV_shapeI_gradshapeJ(i, j) +=
-            swf(fe_vol_data.QPointXYZ(qp)) * fe_vol_data.ShapeValue(i, qp) *
-            fe_vol_data.ShapeGrad(j, qp) * fe_vol_data.JxW(qp); // G-matrix
-
-          IntV_shapeI_shapeJ(i, j) +=
-            swf(fe_vol_data.QPointXYZ(qp)) * fe_vol_data.ShapeValue(i, qp) *
-            fe_vol_data.ShapeValue(j, qp) * fe_vol_data.JxW(qp); // M-matrix
-        }                                                        // for qp
-      }                                                          // for j
-
-      for (const auto& qp : fe_vol_data.GetQuadraturePointIndices())
-      {
-        IntV_shapeI(i) +=
-          swf(fe_vol_data.QPointXYZ(qp)) * fe_vol_data.ShapeValue(i, qp) * fe_vol_data.JxW(qp);
-      } // for qp
-    }   // for i
-
-    //  surface integrals
-    for (size_t f = 0; f < cell_num_faces; ++f)
-    {
-      const auto fe_srf_data = cell_mapping.MakeSurfaceFiniteElementData(f);
-      IntS_shapeI_shapeJ[f] = DenseMatrix<double>(cell_num_nodes, cell_num_nodes, 0.0);
-      IntS_shapeI[f] = Vector<double>(cell_num_nodes, 0.);
-      IntS_shapeI_gradshapeJ[f] = DenseMatrix<Vector3>(cell_num_nodes, cell_num_nodes);
-
-      for (unsigned int i = 0; i < cell_num_nodes; ++i)
-      {
-        for (unsigned int j = 0; j < cell_num_nodes; ++j)
-        {
-          for (const auto& qp : fe_srf_data.GetQuadraturePointIndices())
-          {
-            IntS_shapeI_shapeJ[f](i, j) += swf(fe_srf_data.QPointXYZ(qp)) *
-                                           fe_srf_data.ShapeValue(i, qp) *
-                                           fe_srf_data.ShapeValue(j, qp) * fe_srf_data.JxW(qp);
-            IntS_shapeI_gradshapeJ[f](i, j) += swf(fe_srf_data.QPointXYZ(qp)) *
-                                               fe_srf_data.ShapeValue(i, qp) *
-                                               fe_srf_data.ShapeGrad(j, qp) * fe_srf_data.JxW(qp);
-          } // for qp
-        }   // for j
-
-        for (const auto& qp : fe_srf_data.GetQuadraturePointIndices())
-        {
-          IntS_shapeI[f](i) +=
-            swf(fe_srf_data.QPointXYZ(qp)) * fe_srf_data.ShapeValue(i, qp) * fe_srf_data.JxW(qp);
-        } // for qp
-      }   // for i
-    }     // for f
-
-    return UnitCellMatrices{IntV_gradshapeI_gradshapeJ,
-                            IntV_shapeI_gradshapeJ,
-                            IntV_shapeI_shapeJ,
-                            IntV_shapeI,
-
-                            IntS_shapeI_shapeJ,
-                            IntS_shapeI_gradshapeJ,
-                            IntS_shapeI};
-  };
-
   const size_t num_local_cells = grid_->local_cells.size();
   unit_cell_matrices_.resize(num_local_cells);
 
   for (const auto& cell : grid_->local_cells)
-    unit_cell_matrices_[cell.local_id] = ComputeCellUnitIntegrals(cell, *swf_ptr);
+    unit_cell_matrices_[cell.local_id] =
+      UnitCellMatrices::Compute(sdm, cell, options_.geometry_type);
 
   const auto ghost_ids = grid_->cells.GetGhostGlobalIDs();
   for (uint64_t ghost_id : ghost_ids)
     unit_ghost_cell_matrices_[ghost_id] =
-      ComputeCellUnitIntegrals(grid_->cells[ghost_id], *swf_ptr);
+      UnitCellMatrices::Compute(sdm, grid_->cells[ghost_id], options_.geometry_type);
 
   // Assessing global unit cell matrix storage
   std::array<size_t, 2> num_local_ucms = {unit_cell_matrices_.size(),
