+++ test/python/modules/linear_boltzmann_solvers/dsa/acceleration_diffusion_cfem.cc
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
 // SPDX-License-Identifier: MIT
 
+#include "framework/math/unit_cell_matrices/unit_cell_matrices.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/math/spatial_discretization/finite_element/piecewise_linear/piecewise_linear_continuous.h"
 #include "modules/linear_boltzmann_solvers/lbs_problem/acceleration/acceleration.h"
@@ -51,79 +52,8 @@ acceleration_Diffusion_CFEM(std::shared_ptr<MeshContinuum> grid)
   // Build unit integrals
   for (const auto& cell : grid->local_cells)
   {
-    const auto& cell_mapping = sdm.GetCellMapping(cell);
-    const size_t cell_num_faces = cell.faces.size();
-    const size_t cell_num_nodes = cell_mapping.GetNumNodes();
-    const auto fe_vol_data = cell_mapping.MakeVolumetricFiniteElementData();
-
-    DenseMatrix<double> IntV_gradshapeI_gradshapeJ(cell_num_nodes, cell_num_nodes, 0.0);
-    DenseMatrix<double> IntV_shapeI_shapeJ(cell_num_nodes, cell_num_nodes, 0.0);
-    Vector<double> IntV_shapeI(cell_num_nodes, 0.0);
-
-    std::vector<DenseMatrix<double>> IntS_shapeI_shapeJ(cell_num_faces);
-    std::vector<DenseMatrix<Vector3>> IntS_shapeI_gradshapeJ(cell_num_faces);
-    std::vector<Vector<double>> IntS_shapeI(cell_num_faces);
-
-    // Volume integrals
-    for (unsigned int i = 0; i < cell_num_nodes; ++i)
-    {
-      for (unsigned int j = 0; j < cell_num_nodes; ++j)
-      {
-        for (const auto& qp : fe_vol_data.GetQuadraturePointIndices())
-        {
-          IntV_gradshapeI_gradshapeJ(i, j) +=
-            fe_vol_data.ShapeGrad(i, qp).Dot(fe_vol_data.ShapeGrad(j, qp)) *
-            fe_vol_data.JxW(qp); // K-matrix
-
-          IntV_shapeI_shapeJ(i, j) += fe_vol_data.ShapeValue(i, qp) *
-                                      fe_vol_data.ShapeValue(j, qp) *
-                                      fe_vol_data.JxW(qp); // M-matrix
-        }                                                  // for qp
-      }                                                    // for j
-
-      for (const auto& qp : fe_vol_data.GetQuadraturePointIndices())
-      {
-        IntV_shapeI(i) += fe_vol_data.ShapeValue(i, qp) * fe_vol_data.JxW(qp);
-      } // for qp
-    }   // for i
-
-    //  surface integrals
-    for (size_t f = 0; f < cell_num_faces; ++f)
-    {
-      const auto fe_srf_data = cell_mapping.MakeSurfaceFiniteElementData(f);
-      IntS_shapeI_shapeJ[f] = DenseMatrix<double>(cell_num_nodes, cell_num_nodes, 0.0);
-      IntS_shapeI[f] = Vector<double>(cell_num_nodes, 0.0);
-      IntS_shapeI_gradshapeJ[f] = DenseMatrix<Vector3>(cell_num_nodes, cell_num_nodes);
-
-      for (unsigned int i = 0; i < cell_num_nodes; ++i)
-      {
-        for (unsigned int j = 0; j < cell_num_nodes; ++j)
-        {
-          for (const auto& qp : fe_srf_data.GetQuadraturePointIndices())
-          {
-            IntS_shapeI_shapeJ[f](i, j) +=
-              fe_srf_data.ShapeValue(i, qp) * fe_srf_data.ShapeValue(j, qp) * fe_srf_data.JxW(qp);
-            IntS_shapeI_gradshapeJ[f](i, j) +=
-              fe_srf_data.ShapeValue(i, qp) * fe_srf_data.ShapeGrad(j, qp) * fe_srf_data.JxW(qp);
-          } // for qp
-        }   // for j
-
-        for (const auto& qp : fe_srf_data.GetQuadraturePointIndices())
-        {
-          IntS_shapeI[f](i) += fe_srf_data.ShapeValue(i, qp) * fe_srf_data.JxW(qp);
-        } // for qp
-      }   // for i
-    }     // for f
-
-    unit_cell_matrices[cell.local_id] = UnitCellMatrices{IntV_gradshapeI_gradshapeJ,
-                                                         {},
-                                                         IntV_shapeI_shapeJ,
-                                                         IntV_shapeI,
-
-                                                         IntS_shapeI_shapeJ,
-                                                         IntS_shapeI_gradshapeJ,
-                                                         IntS_shapeI};
-  } // for cell
+    unit_cell_matrices[cell.local_id] = UnitCellMatrices::Compute(sdm, cell);
+  }
 
   // Make solver
   DiffusionPWLCSolver solver("SimTest92b_DSA_PWLC",
